01-ai-agents/my-projects/pyproject.toml
[project]
name = "my-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.27.0",
    "openai-agents>=0.2.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
    "sqlite-vec>=0.1.6",
]
